# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BHTune API",
    "description": "HTTP API for BHTune: DCS/PLC PID-loop templates, and the tune-run history recorded by the CLI and this server.",
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/runs/{id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Fetch one run's full detail.",
        "description": "`GET /api/runs/{id}` -- 404 if no run has that id.",
        "operationId": "show_run",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Run id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The full recorded detail for one run.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "No run with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "runs"
        ],
        "summary": "Delete one run and its recorded samples/results/write-back audit rows.",
        "description": "`DELETE /api/runs/{id}` -- 404 if no run has that id, 409 if the run's own recorded\n`outcome` is still [`TuneOutcome::Running`] (deleting the row out from under an in-flight\ntask would corrupt whatever it tries to write next; cancel it first). Deliberately checks\nthe run row's own `outcome` rather than [`crate::active_run::ActiveRun`]'s in-memory\nactive-run registry entry: `drive()` persists every terminal outcome (`persist_results` then\n`TuneRunRow::complete`/`fail`/`abort`) *before* returning, and `ActiveRun::release` only\nruns strictly after `drive()` returns (see `routes::runs::start_run`'s spawned task), so\nthere is a real -- if brief -- window where a run's outcome is already durably\n`completed`/`failed`/`aborted` but `ActiveRun` hasn't been told the registry entry is free yet.\nChecking the DB's own authoritative, durable state instead of the best-effort in-memory\ntracker closes that race outright, rather than requiring the caller to retry (as\n`frontend/e2e/tune.spec.ts`'s `startTune()` already has to for the equivalent gap on the\n*start* side).",
        "operationId": "delete_run",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Run id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Run deleted."
          },
          "404": {
            "description": "No run with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The run has not finished yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{id}/stream": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Stream per-tick engine state for one run over Server-Sent Events.",
        "description": "`GET /api/runs/{id}/stream` -- 404 if no run has that id. Emits one `initial` event (JSON\nbody: [`InitialReadingsResponse`]) as soon as the driver's initial snapshot is persisted,\nthen a `sample` event (JSON body: [`SampleResponse`], the same shape\n`GET /api/runs/{id}`'s `samples` array already uses) for every tick recorded so far and\nevery new tick recorded while connected, followed by exactly one final `done` event (JSON\nbody: [`RunStreamDone`]) once the run reaches a terminal outcome -- after which the\nconnection closes. Safe to open at any point in a run's lifecycle, including after it has\nalready finished: the initial snapshot (when available) and every sample are replayed once,\nimmediately followed by `done`.",
        "operationId": "stream_run",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Run id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A `text/event-stream` with an optional `initial` event (data: InitialReadingsResponse), `sample` events (data: SampleResponse), and one final `done` event (data: RunStreamDone).",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SampleResponse"
                }
              }
            }
          },
          "404": {
            "description": "No run with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "The JSON body of every non-2xx response: `{\"error\": \"<message>\"}`. `pub`/`ToSchema` so\nevery fallible `#[utoipa::path]` response can reference it (`body = ErrorBody`) and the\ngenerated OpenAPI spec -- and therefore the generated frontend TS client -- accurately\ntypes error bodies instead of `content?: never`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "RunDetailResponse": {
        "type": "object",
        "required": [
          "id",
          "tag_name",
          "driver",
          "outcome",
          "started_at",
          "template_name",
          "template_origin",
          "allow_uncertain_quality",
          "config",
          "pid_parameter_labels",
          "samples",
          "results",
          "writes",
          "mv_actuations"
        ],
        "properties": {
          "allow_uncertain_quality": {
            "type": "boolean",
            "description": "Whether this run accepted `Uncertain` OPC quality, captured when the run started."
          },
          "bridge_host": {
            "type": [
              "string",
              "null"
            ],
            "description": "The resolved bridge host this run actually used, matching `opc_server` above."
          },
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "config": {
            "$ref": "#/components/schemas/LoopConfig"
          },
          "driver": {
            "$ref": "#/components/schemas/TuneDriver"
          },
          "effective_tuning": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EffectiveTuning",
                "description": "Concrete global timing and safety values frozen when this run was prepared. `None`\nidentifies a run created before effective-tuning snapshots were stored."
              }
            ]
          },
          "failure_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "gateway_compatibility": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GatewayCompatibilityResponse",
                "description": "Gateway protocol snapshot recorded before this run's live mutation, when the run used\nOPC DA. `None` for simulator/replay runs and for OPC runs that predate the snapshot."
              }
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "initial_readings": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InitialReadingsResponse"
              }
            ]
          },
          "mv_actuations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MvActuationResponse"
            }
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "opc_server": {
            "type": [
              "string",
              "null"
            ],
            "description": "The resolved OPC DA server ProgID this run actually used, or `None` for a\nsimulator/replay run (`db-run-request-snapshot`). This is what `history revert`\ntrusts over any `--server` flag -- see `bhtune_runtime::history::revert_run`."
          },
          "original_request": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StartRunRequest",
                "description": "This run's own `request_json` (`db-run-request-snapshot`), parsed back into a\n[`StartRunRequest`], or `None` if it isn't usable -- see [`parse_stored_request`].\nPowers the run detail page's \"Duplicate this run\" action (`ui-prefill-last-run`):\nunlike `GET /api/runs/last-request`, which only ever answers for the single newest\nrun, this lets the New tune form seed itself from *this specific* historical run\nregardless of how many later runs exist."
              }
            ]
          },
          "outcome": {
            "$ref": "#/components/schemas/TuneOutcome"
          },
          "pid_constant_tags": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PidConstantTagsResponse",
                "description": "`Some` exactly when `routes::runs::require_writable_run`'s tag-presence check would\npass -- i.e. when all three PID constant tags were configured on this run. The\nfrontend uses this (together with `driver`/`outcome`/`opc_server`/`bridge_host`) to\ndecide whether the post-hoc write/revert buttons are enabled and, when they are not,\nto explain why -- without duplicating `require_writable_run`'s logic client-side or\ndiscovering ineligibility only after a failed request (`api-post-run-write`)."
              }
            ]
          },
          "pid_parameter_labels": {
            "$ref": "#/components/schemas/PidParameterLabelsResponse",
            "description": "Operator-facing calculated-result column labels from the run's historical template\nsnapshot. Empty user-template suffixes use the conventional P/I/D labels."
          },
          "restore_detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "restore_status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RestoreStatus"
              }
            ]
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultResponse"
            }
          },
          "samples": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SampleResponse"
            }
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "tag_name": {
            "type": "string"
          },
          "template_name": {
            "type": "string",
            "description": "Name of the template snapshotted onto this run at start time -- not necessarily what\n`template_name` currently resolves to in the catalog (`safety-run-snapshot`)."
          },
          "template_origin": {
            "$ref": "#/components/schemas/TemplateOrigin"
          },
          "timing_metrics": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TimingMetrics"
              }
            ]
          },
          "writes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WriteResponse"
            }
          }
        }
      },
...
      "SampleResponse": {
        "type": "object",
        "description": "One recorded tick: the [`Tick`] input and resulting engine state, plus the driver-\nreported PV quality at read time. `Tick`/`MrftState` already derive `Serialize` in\n`bhtune-core` (they round-trip through golden-trace fixtures too), so they're embedded\ndirectly rather than re-projected field-by-field like the other DTOs here.",
        "required": [
          "tick_index",
          "sample",
          "state",
          "pv_quality"
        ],
        "properties": {
          "pv_quality": {
            "$ref": "#/components/schemas/SampleQuality"
          },
          "sample": {
            "$ref": "#/components/schemas/Tick"
          },
          "state": {
            "$ref": "#/components/schemas/MrftState"
          },
          "tick_index": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "TuneDriver": {
        "type": "string",
        "description": "Which [`crate`]-agnostic I/O driver a run used. Lives in `bhtune-db` rather than\n`bhtune-core` because it's a persistence/orchestration concept (which adapter drove this\nrun), not a domain concept the pure MRFT engine itself needs to know about.",
        "enum": [
          "opcda",
          "simulator",
          "replay"
        ]
      },
      "TuneOutcome": {
        "type": "string",
        "description": "A run's lifecycle state.",
        "enum": [
          "running",
          "completed",
          "failed",
          "aborted"
        ]
      },
...
    }
  },
...
}